+++ scenarioloader.py
@@ -67,7 +67,7 @@ class ScenarioLoader(object):
 						v_z = float(parts[9])*progaconstants.FOOT2MT
 						pitch = float(parts[4])
 						bank = float(parts[5])
-						onground = float(parts[7])
+						onground = int(parts[7])
 						airspeed = float(parts[19])
 	
 						track.addStep(simulation_timestamp, lat, lon, altitude, v_x, v_y, v_z, heading, pitch, bank, onground, airspeed)
